Removed sampling module
