+++ app/src/smf_ecg_bioz.c
@@ -5,7 +5,7 @@
 #include <zephyr/zbus/zbus.h>
 #include <zephyr/drivers/sensor.h>
 
-#include "sampling_module.h"
+#include "hpi_common_types.h"
 #include "hw_module.h"
 #include "ui/move_ui.h"
