+++ app/src/ui/hp_ui_common.c
@@ -15,7 +15,7 @@
 
 #include "hw_module.h"
 #include "power_ctrl.h"
-#include "sampling_module.h"
+#include "hpi_common_types.h"
 #include "ui/move_ui.h"
 
 #include <display_sh8601.h>
