+++ app/src/ui/screens/scr_bpt.c
@@ -8,7 +8,7 @@
 #include <app_version.h>
 #include <zephyr/logging/log.h>
 
-#include "sampling_module.h"
+#include "hpi_common_types.h"
 #include "hw_module.h"
 
 #include "ui/move_ui.h"
