+++ app/src/ui/screens/scr_clock.c
@@ -11,7 +11,7 @@
 #include <zephyr/logging/log.h>
 #include <zephyr/drivers/rtc.h>
 
-#include "sampling_module.h"
+#include "hpi_common_types.h"
 
 #include "ui/move_ui.h"
